Extract ExampleList
