+++ backend/api/views/__init__.py
@@ -1,10 +1,9 @@
 from .annotation import *
 from .auto_labeling import *
 from .comment import *
-from .document import *
+from .example import *
 from .export_dataset import *
 from .health import *
-from .image import *
 from .import_dataset import *
 from .import_export import *
 from .label import *
